agents/data_ingestor/requirements.txt
google-cloud-storage>=2.16.0
google-cloud-pubsub>=2.21.0
